vertical-timeline & dropdownmenu
Previously when I committed something using Github Desktop the old commit message and additional comment would be gone, and those fields would have been empty. But after the latest stable update, due to some reason that is not happening anymore. The commit message and the additional comment still remains. Although it is not a problem, that we really should be concerned about but it should be solved. That's what I think.